# docs/swagger.yaml
basePath: /
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  response.Response:
    properties:
      body: {}
      status:
        type: integer
    type: object
  response.UserGenderSwag:
    enum:
    - Male
    - Female
    - Other
    type: string
    x-enum-varnames:
    - Male
    - Female
    - Other
  response.UserGoogleSwag:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      description:
        type: string
      firstname:
        type: string
      gender:
        $ref: '#/definitions/response.UserGenderSwag'
      id:
        type: string
      login:
        type: string
      middlename:
        type: string
      phonenumber:
        type: string
      surname:
        type: string
    type: object
...
host: mailhub.su
info:
  contact: {}
  description: API server for MailHub
  title: API MailHub
  version: "1.0"
paths:
...
  /api/v1/auth/gAuth:
    get:
      consumes:
      - application/json
      description: GoogleAuth Handles user.
      parameters:
      - description: code from oauth
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Auth successful
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not fount
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to create session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GoogleAuth User
      tags:
      - auth-gmail
  /api/v1/auth/getAuthURL:
    get:
      consumes:
      - application/json
      description: GetAuthURL Handles url.
      produces:
      - application/json
      responses:
        "200":
          description: Auth successful
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to create url
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetAuthURL
      tags:
      - auth-gmail
...
  /api/v1/auth/signupGMailUser:
    post:
      consumes:
      - application/json
      description: Handles user signup Gmail.
      parameters:
      - description: New user details for signup
        in: body
        name: newUser
        required: true
        schema:
          $ref: '#/definitions/response.UserGoogleSwag'
      produces:
      - application/json
      responses:
        "200":
          description: Signup successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to add user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: User signup Gmail
      tags:
      - auth-gmail
...
swagger: "2.0"